Backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
bitarray==3.1.1
certifi==2025.1.31
charset-normalizer==3.4.1
ckzg==2.1.0
cytoolz==1.0.1
Django==5.1.7
djangorestframework==3.15.2
eth-account==0.13.5
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.6.1
eth-rlp==2.2.0
eth-typing==5.2.0
eth-utils==5.2.0
eth_abi==5.2.0
frozenlist==1.5.0
hexbytes==1.3.0
idna==3.10
multidict==6.1.0
parsimonious==0.10.0
propcache==0.3.0
pycryptodome==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
pyunormalize==16.0.0
regex==2024.11.6
requests==2.32.3
rlp==4.1.0
sqlparse==0.5.3
toolz==1.0.0
types-requests==2.32.0.20250306
typing_extensions==4.12.2
urllib3==2.3.0
web3==7.9.0
websockets==13.1
yarl==1.18.3
